HIDE - Move enumerable to own package
